+++ packages/material-ui-lab/src/TreeItem/index.d.ts
@@ -1,3 +1,4 @@
 export { default } from './TreeItem';
 export * from './TreeItem';
+export * from './TreeItemContent';
 export { default as useTreeItem } from './useTreeItem';
